+++ tests/ats/test_fixtures.py
@@ -21,6 +21,11 @@ def fixtures(kube_cluster: Cluster):
     ret = kube_cluster.kubectl("apply", filename="cluster-crd.yaml", output_format="json")
     LOGGER.debug("Created cluster CRD")
 
+    # Organization CRD
+    LOGGER.info("Create Organization CRD")
+    ret = kube_cluster.kubectl("apply", filename="https://raw.githubusercontent.com/giantswarm/organization-operator/main/config/crd/security.giantswarm.io_organizations.yaml", output_format="json")
+    LOGGER.debug(f"Created Organization CRD: {ret}")
+
     # Kyverno
     LOGGER.info(f"Install Kyverno {KYVERNO_VERSION}")
     ret = kube_cluster.kubectl("apply --server-side", filename=f"https://github.com/kyverno/kyverno/releases/download/{KYVERNO_VERSION}/install.yaml", output_format="json" )
@@ -30,7 +35,21 @@ def fixtures(kube_cluster: Cluster):
     # Our policies
     LOGGER.info("Deploy Kyverno policies for the service-priority cluster label")
     ret = kube_cluster.kubectl("apply", filename="../../policies/ux/clusters-label-service-priority.yaml", output_format="json")
+    ret = kube_cluster.kubectl("apply", filename="../../policies/ux/organization-deletion-when-has-clusters.yaml", output_format="json")
     LOGGER.debug(f"Created kyverno policies result: {ret}")
+
+    # Create Organization namespace
+    LOGGER.info("Create namespace named 'org-giantswarm'")
+    ret = kube_cluster.kubectl("apply", filename="test-organization.yaml", output_format="json")
+    ret = kube_cluster.kubectl("apply", filename="test-empty-organization.yaml", output_format="json")
+    LOGGER.debug(f"Created giantswarm organization and its namespace: {ret}")
+
+    # This block is commented because k8s v1.24 is needed to use '--subresource', and at the time of writing this, dats/dabs is broken and we can't specify which version to use.
+    # Patch Organizations so that they contain their namespace on their status field, like organization-operator does
+    # ret = kube_cluster.kubectl(f"patch organization giantswarm --subresource status --patch-file organization-giantswarm-patch.yaml")
+    # LOGGER.debug(f"Patched giantswarm organization status with namespace: {ret}")
+    # ret = kube_cluster.kubectl(f"patch organization empty --subresource status --patch-file organization-empty-patch.yaml")
+    # LOGGER.debug(f"Patched empty organization status with namespace: {ret}")
 
     # Test cluster CR
     LOGGER.info("Create cluster.x-k8s.io/v1beta1 named 'test-cluster'")